Added the total time functionality
