+++ src/store/actions/tutorialsActions.js
@@ -2,6 +2,7 @@ import * as actions from "./actionTypes";
 import Elasticlunr from "../../helpers/elasticlunr";
 import { checkOrgHandleExists, checkUserHandleExists } from "./authActions";
 import _ from "lodash";
+import { updateTotalTime } from "../../helpers/updateTotalTime";
 
 const tutorials_index = new Elasticlunr(
   "tutorial_id",
@@ -175,23 +176,7 @@ export const createTutorial =
     }
   };
 
-const updateTotalTime = async (firestore,tutorial_id, time) => {
-  try {
-    const tutorialDoc = await firestore
-      .collection("tutorials")
-      .doc(tutorial_id)
-      .get();
-    const currentTotalTime = tutorialDoc.get("total_time") || 0;
-    const newTotalTime = parseInt(currentTotalTime) + parseInt(time);
 
-    await firestore.collection("tutorials").doc(tutorial_id).update({
-      total_time: newTotalTime,
-      updatedAt: firestore.FieldValue.serverTimestamp()
-    });
-  } catch (e) {
-    console.log(e.message);
-  }
-};
 
 const checkUserOrOrgHandle = handle => async firestore => {
   const userHandleExists = await checkUserHandleExists(handle)(firestore);
